Moved `WavelengthCalibration._evaluate_solution` to `core` as `evaluate_wavelength_solution`
