+++ goodman_pipeline/core/__init__.py
@@ -27,6 +27,7 @@ from .core import (astroscrappy_lacosmic,
                    extraction,
                    extract_fractional_pixel,
                    extract_optimal,
+                   evaluate_wavelength_solution,
                    fix_keywords,
                    fractional_sum,
                    get_best_flat,
